OSO-49 moved logger mock to TestMocks in order to reduce code duplication
